Tag bitmask version 0.3.0

Conflicts:
	.coveragerc
	.tx/config
	NEWS.rst
	setup.cfg
	src/leap/base/checks.py
	src/leap/base/exceptions.py
	src/leap/base/tests/test_checks.py
	src/leap/eip/checks.py
	src/leap/gui/firstrun/login.py
	src/leap/gui/locale_rc.py
	src/leap/util/polkit.py
	tests/test_qt_environment.py
	tox.ini

+++ pkg/postmkvenv.sh
@@ -9,14 +9,14 @@
 # use import PyQt4; PyQt4.__path__ instead
 
 platform='unknown'
-unamestr=`uname`
+unamestr=$(uname)
 if [[ "$unamestr" == 'Linux' ]]; then
    platform='linux'
 elif [[ "$unamestr" == 'Darwin' ]]; then
    platform='darwin'
 fi
 
-LIBS=( PyQt4 sip.so )
+LIBS=( PySide pysideuic )
 
 PYTHON_VERSION=python$(python -c "import sys; print (str(sys.version_info[0])+'.'+str(sys.version_info[1]))")
 VAR=( $(which -a $PYTHON_VERSION) )
@@ -34,5 +34,7 @@ fi
 
 for LIB in ${LIBS[@]}
 do
-    ln -s $LIB_SYSTEM_PATH/$LIB $LIB_VIRTUALENV_PATH/$LIB 
+    if [[ ! -e $LIB_VIRTUALENV_PATH/$LIB ]]; then
+       ln -s $LIB_SYSTEM_PATH/$LIB $LIB_VIRTUALENV_PATH/$LIB
+    fi
 done
